model1: do not write performance related summary fields
this is because doing so will cause any run to fail once these fields
are actually supported by `flow`.

Note: I'm not 100% sure what ELAPSED is supposed to write. Wall time
or simulation time? Anyway it does not make the test much less useful
if that field is not written.