OPERATE Keyword Tests
A series of tests to test the OPERATE keyword, currently only for the EDIT section.
The tests are designed around a "base case" and are generated in such a manner that the results should be identical, or very similar, to the "base case".

The tests confirm user feedback that the DEPTH, TRANX/Y/Z options are not working.