SPE10 Tests Update - WIP
Update SPE10 cases for model 1 tests. Fixed typos in original cases and created two-phase cases. All four cases now give virtually the same results, and the cumulative oil is now similar to the published cases of around 45 Mstb.

Wasted time on trying to get the commercial simulator to run the first two cases to no available, as the the producer gets shut-in at around 2,000 days.

I have placed the cases in a separate directory, and will also add the model two cases in this directory as well when there are ready.

See SPE10.md for case description and commentary.