Use Markdown to Document Results
Here we generate the results in OPM ResInsight via PNG files and then use markdown to organise the results. The intent, if everyone agrees, is to replace the the ODT file with this PR.

I have placed all the plots in the one markdown file, but I could also place each case in a separate file, if that is less complicated to review the results.

Note that there a few limitations in OPM ResInsight at the moment that prevents the generation of the plots being fully automatic. Nevertheless, once the plots are setup generating a new version based on the master being updated is virtually automatic.

Comments appreciated.