Add test case for directional mobilities
Add new test case for solvent and directional mobilities.
This is the same test case as SPE1CASE2_SOLVENT.DATA except
that a second SATNUM region has been added by copying the original
saturation tables and modifying the saturation table in SOF3
slightly for the second region. The relative permeabilities
in the x- and y-directions then use this table instead of the
original table.